requirements.txt
Flask
gunicorn
numpy
opencv-python
Pillow
torch
torchvision
streamlit
